+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- [#1934](https://github.com/teambit/bit/issues/1934) fix capsule compilers when configured to work with custom files
 - [#1885](https://github.com/teambit/bit/issues/1885) introduce new flags `--dependents` and `--dependencies` for `bit import` to import them all directly
 - [#1924](https://github.com/teambit/bit/issues/1924) avoid generating old dists files when using capsule compilers